// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/dispensary/dispensary-main/sales-main/sales-return/model/invoice-item-detail-tobe-return.model.ts
export class InvoiceItemDetailToBeReturn {
    ItemId: number = 0;
    ItemName: string = '';
    BatchNo: string = '';
    SalePrice: number = 0;
    Quantity: number = 0;
    SoldQty: number = 0;
    ReturnedQty: number = 0;
    PreviouslyReturnedQty: number = 0;
    AvailableQty: number = 0;
    SubTotal: number = 0;
    TaxableAmount: number = 0;
    DiscountAmount: number = 0;
    DiscountPercentage: number = 0;
    VATPercentage: number = 0;
    VATAmount: number = 0;
    TotalAmount: number = 0;
    PriceCategoryId: number = 0;
    BillNo: string = '';
    InvoiceId: number = 0;
    InvoiceItemId: number = 0;
    FiscalYearName: string = '';
    FiscalYearId: number = 0;
    Checked: boolean = false;
    InvoiceReturnId: boolean = false;
    InvoiceNo: number = 0;
    CreatedOn: string = "";
    IsCoPayment: boolean = false;
    ReturnCashAmount: number = 0;
    ReturnCreditAmount: number = 0;
    PatientId: number = 0;
    SchemeId: number = 0;


}